requirements.txt
keyboard==0.13.5
openai_whisper==20230314
PyAudio==0.2.13
pynput==1.7.6
sounddevice==0.4.6
torch==2.0.0
transformers==4.28.0
TTS==0.13.3
